requirements.txt
streamlit
transformers
torch
zipfile36
python-docx
PyPDF2
datasets 
scikit-learn
accelerate